refactor(architecture): reorganize project structure to align with Clean Architecture principles
- Added `handleError` utility to the application layer for centralized error handling.
- Updated use cases to use the new error handling approach.
- Refactored controllers to decouple them from the use cases and improve separation of concerns.
- Created `IUserRepository` interface to abstract the repository layer and adhere to Dependency Inversion Principle.
- Updated tests to match the new architecture.

This refactor improves separation of concerns, makes the codebase more maintainable, and ensures consistency across layers.
