--- backend-app/src/config/env.js
@@ -1,6 +1,0 @@
-require("dotenv").config();
-
-module.exports = {
-  PORT: process.env.PORT,
-  MONGO_URI: process.env.MONGO_URI,
-};
