+++ backend-app/src/domain/interfaces/userRepository.js
@@ -1,0 +1,25 @@
+module.exports = class IUserRepository {
+    async findByEmail(email) {
+      throw new Error("Method not implemented");
+    }
+  
+    async create(user) {
+      throw new Error("Method not implemented");
+    }
+  
+    async update(userId, userData) {
+      throw new Error("Method not implemented");
+    }
+  
+    async findAll() {
+      throw new Error("Method not implemented");
+    }
+  
+    async findById(userId) {
+      throw new Error("Method not implemented");
+    }
+  
+    async delete(userId) {
+      throw new Error("Method not implemented");
+    }
+  };
