+++ backend-app/src/infrastructure/config/env.js
@@ -1,0 +1,8 @@
+require("dotenv").config();
+
+module.exports = {
+  PORT: process.env.PORT,
+  MONGO_URI: process.env.MONGO_URI,
+  JWT_SECRET: process.env.JWT_SECRET
+
+};
