+++ backend-app/src/infrastructure/database/mongoose/mongoose.js
@@ -1,0 +1,13 @@
+const mongoose = require("mongoose");
+const config = require("../../config/env");
+const connectDB = async () => {
+  try {
+    await mongoose.connect(config.MONGO_URI);
+    console.log("MongoDB connected");
+  } catch (error) {
+    console.error("MongoDB connection error:", error);
+    process.exit(1);
+  }
+};
+
+module.exports = connectDB;
