+++ backend-app/src/interface/swagger/swagger.js
@@ -1,0 +1,15 @@
+const swaggerJsDoc = require("swagger-jsdoc");
+
+const options = {
+  definition: {
+    openapi: "3.0.0",
+    info: {
+      title: "User API",
+      version: "1.0.0",
+      description: "API for managing users",
+    },
+  },
+  apis: ["./src/interface/routes/*.js"], 
+};
+
+module.exports = swaggerJsDoc(options);
