--- backend-app/src/interfaces/middleware/authMiddleware.js
@@ -1,18 +1,0 @@
-const jwt = require("jsonwebtoken");
-require("dotenv").config(); 
-
-exports.authenticate = (req, res, next) => {
-  const token = req.header("Authorization")?.replace("Bearer ", "");
-
-  if (!token) {
-    return res.status(401).json({ error: "Access denied. No token provided." });
-  }
-
-  try {
-    const decoded = jwt.verify(token, process.env.JWT_SECRET); 
-    req.user = decoded;
-    next();
-  } catch (error) {
-    res.status(400).json({ error: "Invalid token." });
-  }
-};
