+++ backend-app/src/usecases/errors/duplicateUserError.js
@@ -1,0 +1,8 @@
+class DuplicateUserError extends Error {
+    constructor(message) {
+      super(message);
+      this.name = "DuplicateUserError";
+    }
+  }
+  
+  module.exports = DuplicateUserError;
