+++ backend-app/src/usecases/errors/index.js
@@ -1,0 +1,7 @@
+const NotFoundError = require("./notFoundError");
+const ValidationError = require("./validationError");
+
+module.exports = {
+  NotFoundError,
+  ValidationError,
+};
