+++ backend-app/src/usecases/errors/notFoundError.js
@@ -1,0 +1,8 @@
+class NotFoundError extends Error {
+    constructor(message) {
+      super(message);
+      this.name = "NotFoundError";
+    }
+  }
+  
+  module.exports = NotFoundError;
