+++ backend-app/src/usecases/errors/validationError.js
@@ -1,0 +1,8 @@
+class ValidationError extends Error {
+    constructor(message) {
+      super(message);
+      this.name = "ValidationError";
+    }
+  }
+  
+  module.exports = ValidationError;
