+++ backend-app/src/usecases/user/getAllUsers.js
@@ -1,5 +1,10 @@
+const IUserRepository = require("../../domain/interfaces/userRepository");
+
 class GetAllUsers {
     constructor(userRepository) {
+      if (!(userRepository instanceof IUserRepository)) {
+        throw new Error("Invalid user repository: Must implement IUserRepository");
+      }
       this.userRepository = userRepository;
     }
